Rework and addition of tests for dat functions
This started out as a project to replace the use of
the expand() function when generating new variable names
the simple uses of sub().

Since that entailed so many changes I was motivated to
create some tests.  So now there exists a simple test of
almost all dat functions, using a little 11 record, 5 station,
extract of a VERTEX netcdf file.

The only missing simple tests are for the dat.S* functions which
require 3rd moments, and for krypton data, which doesn't
exist in the VERTEX data.

Running these tests, as expected, exposed quite a few bugs.