Add workarounds when netcdf attributes are missing
When assembling a list of time series variables from a set of files
generate the R objects from the variable by name in the files with
the largest number of attributes. So if a short_name attribute
is missing in some files, and present in others, the search by
short_name should work.

Handle situation of missing _FillValue attibute, and the situation
where the _FillValue attribute may have been added after the
data was written.