More updates to nts indexing.
In general, with well behaved time series, there shouldn't be any
noticeble changes.

For y <- x[i], where x and i are nts: first match times between x and i.
If i is logical, then i and x must have the same number of columns,
and return matching times where i is TRUE.  If i is not logical, return
matching times, ignoring contents of i.  Result is not an nts.
In the previous version, matching of times was not done, which works fine
if x and i have the same times.

For y <- x[i,], where x and i are nts: first match times between x and i.
Then if i is logical then i must have only one column, and values with
matching times for which i is TRUE are returned. If i is not logical,
return rows with matching times, basically treating i as all TRUE.
Result is an nts.
In the previous version, matching of times was not done, which works
if x and i have the same times. This should be more general, if a little slower.

For x[i] <- val, where x and i are nts: first match times between x and i.
If i is logical, then i and x must have the same number of columns.
Matching times where i is TRUE are replaced with val. If i is not logical,
replace those values with matching times, basically treating i as all TRUE.
In the previous version, there was no requirement on the number of columns.
If the RHS is a vector or matrix, it can be tricky to avoid the error:
"number of items to replace is not a multiple of replacement length",
unless you're using the same logical expression on the RHS and LHS, such as:
x[y > 5] <- z[y > 5]

For x[i,] <- val, where x and i are nts: first match times between x and i.
If i is logical, then i must have only one column, and values with
matching times for which i is TRUE are replaced.
If i is not logical, replace those values with matching times, basically
treating i as all TRUE.
Again, if the RHS is a vector or matrix, it can be tricky to avoid the
replacement length error, unless you're using the same logical expression
on the RHS and LHS, such as:
x[y[,3] > 5,] <- z[y[,3] > 5,]

Add help examples and tests for indexing.

In match.nts, set the default dt.eps to 1/10 of the deltat, rather than
1/100.

When indexing x[i...] and matching times of i in x don't return duplicate
times in x, which would happen if dt is larger than some deltas in i.
Add dupOK argument to match.nts, rather than using duplicate() function.